// frontend-mobile/src/lib/types.ts
export interface Message {
  id: string;
  type: 'SMS' | 'Email';
  recipient: string;
  recipientEmail?: string;
  recipientPhone?: string;
  subject?: string;
  content: string;
  status: 'Sent' | 'Scheduled' | 'Failed';
  sentBy: string;
  date: string;
  time?: string;
}